@@ -1,15 +1,17 @@
+import org.ajoberstar.grgit.Grgit
+
 apply from: 'config/keystore.gradle'
 
 buildscript {
     repositories {
+        maven { url "https://plugins.gradle.org/m2/" }
         jcenter()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.0'
-        classpath 'org.ajoberstar:gradle-git:1.7.2'
+        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'org.ajoberstar.grgit:grgit-gradle:4.1.0'
+        classpath 'org.jacoco:org.jacoco.core:0.8.6'
     }
 }
 
@@ -24,7 +26,7 @@ subprojects {
 ext.verifyKeystore()
 
 try {
-    def grgit = org.ajoberstar.grgit.Grgit.open(dir: '.')
+    def grgit = Grgit.open(currentDir: project.rootDir)
     def lastCommit = grgit.head()
 
     project.ext.versionNum = grgit.log(includes:['HEAD']).size()
@@ -33,16 +35,31 @@ try {
     if (project.ext.versionName == null) {
         project.ext.versionName = 'DEV'
     }
-} catch (Exception ex) {
-    project.ext.versionNum = 0
+} catch (Exception ignored) {
+    project.ext.versionNum = 1
     project.ext.versionName = 'DEV'
     project.ext.versionDate = new Date()
 }
 
-project.ext.minSdkVersion = 16
-project.ext.compileSdkVersion = 27
-project.ext.buildToolsVersion = '27.0.3'
-project.ext.supportLibVersion = '27.1.1'
+project.ext {
+    minSdkVersion = 16
+    compileSdkVersion = 30
+    buildToolsVersion = '30.0.0'
+
+    googleVersions = [
+        glide                 : '4.12.0',
+        material              : '1.3.0',
+    ]
+
+    androidXVersions = [
+        annotation            : '1.1.0',
+        appCompat             : '1.2.0',
+        browser               : '1.3.0',
+    ]
+
+    okioVersion = '2.10.0'
+    jodaVersion = '2.10.10'
+}
 
 task showVersion {
     doLast {
