@@ -14,7 +14,7 @@
 
 package net.openid.appauth;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import net.openid.appauth.browser.AnyBrowserMatcher;
 import net.openid.appauth.browser.BrowserMatcher;
@@ -40,11 +40,19 @@ public class AppAuthConfiguration {
     @NonNull
     private final ConnectionBuilder mConnectionBuilder;
 
+    private final boolean mSkipIssuerHttpsCheck;
+
+    private final boolean mSkipNonceVerification;
+
     private AppAuthConfiguration(
             @NonNull BrowserMatcher browserMatcher,
-            @NonNull ConnectionBuilder connectionBuilder) {
+            @NonNull ConnectionBuilder connectionBuilder,
+            Boolean skipIssuerHttpsCheck,
+            Boolean skipNonceVerification) {
         mBrowserMatcher = browserMatcher;
         mConnectionBuilder = connectionBuilder;
+        mSkipIssuerHttpsCheck = skipIssuerHttpsCheck;
+        mSkipNonceVerification = skipNonceVerification;
     }
 
     /**
@@ -65,12 +73,30 @@ public class AppAuthConfiguration {
     }
 
     /**
+     * Returns <code>true</code> if issuer https validation is disabled, otherwise
+     * <code>false</code>.
+     *
+     * @see Builder#setSkipIssuerHttpsCheck(Boolean)
+     */
+    public boolean getSkipIssuerHttpsCheck() { return mSkipIssuerHttpsCheck; }
+
+    /**
+     * Returns <code>true</code> if nonce verification on response is disabled, otherwise
+     * <code>false</code>.
+     *
+     * @see Builder#setSkipNonceVerification(Boolean)
+     */
+    public boolean getSkipNonceVerification() { return mSkipNonceVerification; }
+
+    /**
      * Creates {@link AppAuthConfiguration} instances.
      */
     public static class Builder {
 
         private BrowserMatcher mBrowserMatcher = AnyBrowserMatcher.INSTANCE;
         private ConnectionBuilder mConnectionBuilder = DefaultConnectionBuilder.INSTANCE;
+        private boolean mSkipIssuerHttpsCheck;
+        private boolean mSkipNonceVerification;
 
         /**
          * Specify the browser matcher to use, which controls the browsers that can be used
@@ -95,11 +121,42 @@ public class AppAuthConfiguration {
         }
 
         /**
+         * Disables https validation for the issuer identifier.
+         *
+         * <p>NOTE: Disabling issuer https validation implies the app is running against an
+         * insecure environment. Enabling this option is only recommended for testing purposes.
+         */
+        public Builder setSkipIssuerHttpsCheck(Boolean skipIssuerHttpsCheck) {
+            mSkipIssuerHttpsCheck = skipIssuerHttpsCheck;
+            return this;
+        }
+
+        /**
+         * Disables nonce verification for value sent in the Authentication Request.
+         *
+         * <p>NOTE: Some Authorization Servers do not return the requested nonce value thus failing
+         * ID token validation. Please consider raising an issue with your Identity Provider and
+         * disabling this option once it is fixed.
+         *
+         * <p>Alternatively, you may avoid sending a nonce by passing <code>null</code> to
+         * {@link AuthorizationRequest.Builder#setNonce}
+         */
+        public Builder setSkipNonceVerification(Boolean skipNonceVerification) {
+            mSkipNonceVerification = skipNonceVerification;
+            return this;
+        }
+
+        /**
          * Creates the instance from the configured properties.
          */
         @NonNull
         public AppAuthConfiguration build() {
-            return new AppAuthConfiguration(mBrowserMatcher, mConnectionBuilder);
+            return new AppAuthConfiguration(
+                mBrowserMatcher,
+                mConnectionBuilder,
+                mSkipIssuerHttpsCheck,
+                mSkipNonceVerification
+            );
         }
 
 
