@@ -18,14 +18,14 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.WorkerThread;
-import android.support.customtabs.CustomTabsCallback;
-import android.support.customtabs.CustomTabsClient;
-import android.support.customtabs.CustomTabsIntent;
-import android.support.customtabs.CustomTabsServiceConnection;
-import android.support.customtabs.CustomTabsSession;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.WorkerThread;
+import androidx.browser.customtabs.CustomTabsCallback;
+import androidx.browser.customtabs.CustomTabsClient;
+import androidx.browser.customtabs.CustomTabsIntent;
+import androidx.browser.customtabs.CustomTabsServiceConnection;
+import androidx.browser.customtabs.CustomTabsSession;
 
 import net.openid.appauth.internal.Logger;
 import net.openid.appauth.internal.UriUtil;
@@ -103,7 +103,7 @@ public class CustomTabManager {
     }
 
     /**
-     * Creates a {@link android.support.customtabs.CustomTabsIntent.Builder custom tab builder},
+     * Creates a {@link androidx.browser.customtabs.CustomTabsIntent.Builder custom tab builder},
      * with an optional list of optional URIs that may be requested. The URI list
      * should be ordered such that the most likely URI to be requested is first. If the selected
      * browser does not support custom tabs, then the URI list has no effect.
@@ -129,7 +129,7 @@ public class CustomTabManager {
     }
 
     /**
-     * Creates a {@link android.support.customtabs.CustomTabsSession custom tab session} for
+     * Creates a {@link androidx.browser.customtabs.CustomTabsSession custom tab session} for
      * use with a custom tab intent, with optional callbacks and optional list of URIs that may
      * be requested. The URI list should be ordered such that the most likely URI to be requested
      * is first. If no custom tab supporting browser is available, this will return {@code null}.
