@@ -14,6 +14,8 @@
 
 package net.openid.appauth;
 
+import static android.os.Looper.getMainLooper;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -21,37 +23,42 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.robolectric.Shadows.shadowOf;
 
 import android.net.Uri;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 import net.openid.appauth.AuthorizationException.GeneralErrors;
 import net.openid.appauth.connectivity.ConnectionBuilder;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
+import org.robolectric.android.util.concurrent.PausedExecutorService;
 import org.robolectric.annotation.Config;
+import org.robolectric.annotation.LooperMode;
+import org.robolectric.shadows.ShadowPausedAsyncTask;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk=16)
+@Config(sdk = 16)
+@LooperMode(LooperMode.Mode.PAUSED)
 public class AuthorizationServiceConfigurationTest {
     private static final int CALLBACK_TIMEOUT_MILLIS = 1000;
     private static final String TEST_NAME = "test_name";
     private static final String TEST_ISSUER = "test_issuer";
     private static final String TEST_AUTH_ENDPOINT = "https://test.openid.com/o/oauth/auth";
     private static final String TEST_TOKEN_ENDPOINT = "https://test.openid.com/o/oauth/token";
+    private static final String TEST_END_SESSION_ENDPOINT = "https://test.openid.com/o/oauth/logout";
     private static final String TEST_REGISTRATION_ENDPOINT = "https://test.openid.com/o/oauth/registration";
     private static final String TEST_USERINFO_ENDPOINT = "https://test.openid.com/o/oauth/userinfo";
     private static final String TEST_JWKS_URI = "https://test.openid.com/o/oauth/jwks";
@@ -69,6 +76,7 @@ public class AuthorizationServiceConfigurationTest {
             + " \"authorization_endpoint\": \"" + TEST_AUTH_ENDPOINT + "\",\n"
             + " \"token_endpoint\": \"" + TEST_TOKEN_ENDPOINT + "\",\n"
             + " \"registration_endpoint\": \"" + TEST_REGISTRATION_ENDPOINT + "\",\n"
+            + " \"end_session_endpoint\": \"" + TEST_END_SESSION_ENDPOINT + "\",\n"
             + " \"userinfo_endpoint\": \"" + TEST_USERINFO_ENDPOINT + "\",\n"
             + " \"jwks_uri\": \"" + TEST_JWKS_URI + "\",\n"
             + " \"response_types_supported\": " + toJson(TEST_RESPONSE_TYPE_SUPPORTED) + ",\n"
@@ -104,20 +112,31 @@ public class AuthorizationServiceConfigurationTest {
             + " \"userinfo_endpoint\": \"" + TEST_USERINFO_ENDPOINT + "\"\n"
             + "}";
 
+    private AutoCloseable mMockitoCloseable;
     private AuthorizationServiceConfiguration mConfig;
     private RetrievalCallback mCallback;
     @Mock HttpURLConnection mHttpConnection;
     @Mock ConnectionBuilder mConnectionBuilder;
+    private PausedExecutorService mPausedExecutorService;
 
     @Before
     public void setUp() throws Exception {
-        MockitoAnnotations.initMocks(this);
+        mMockitoCloseable = MockitoAnnotations.openMocks(this);
         mCallback = new RetrievalCallback();
         mConfig = new AuthorizationServiceConfiguration(
                 Uri.parse(TEST_AUTH_ENDPOINT),
                 Uri.parse(TEST_TOKEN_ENDPOINT),
-                Uri.parse(TEST_REGISTRATION_ENDPOINT));
+                Uri.parse(TEST_REGISTRATION_ENDPOINT),
+                Uri.parse(TEST_END_SESSION_ENDPOINT));
         when(mConnectionBuilder.openConnection(any(Uri.class))).thenReturn(mHttpConnection);
+
+        mPausedExecutorService = new PausedExecutorService();
+        ShadowPausedAsyncTask.overrideExecutor(mPausedExecutorService);
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        mMockitoCloseable.close();
     }
 
     @Test
@@ -137,11 +156,26 @@ public class AuthorizationServiceConfigurationTest {
         AuthorizationServiceConfiguration config = new AuthorizationServiceConfiguration(
                 Uri.parse(TEST_AUTH_ENDPOINT),
                 Uri.parse(TEST_TOKEN_ENDPOINT),
-                null);
+                null,
+                Uri.parse(TEST_END_SESSION_ENDPOINT));
         AuthorizationServiceConfiguration deserialized = AuthorizationServiceConfiguration
                 .fromJson(config.toJson());
         assertThat(deserialized.authorizationEndpoint).isEqualTo(config.authorizationEndpoint);
         assertThat(deserialized.tokenEndpoint).isEqualTo(config.tokenEndpoint);
+        assertThat(deserialized.registrationEndpoint).isNull();
+        assertThat(deserialized.endSessionEndpoint).isEqualTo(config.endSessionEndpoint);
+    }
+
+    @Test
+    public void testSerializationWithoutRegistrationEndpointAndEndSessionEndpoint() throws Exception {
+        AuthorizationServiceConfiguration config = new AuthorizationServiceConfiguration(
+            Uri.parse(TEST_AUTH_ENDPOINT),
+            Uri.parse(TEST_TOKEN_ENDPOINT));
+        AuthorizationServiceConfiguration deserialized = AuthorizationServiceConfiguration
+            .fromJson(config.toJson());
+        assertThat(deserialized.authorizationEndpoint).isEqualTo(config.authorizationEndpoint);
+        assertThat(deserialized.tokenEndpoint).isEqualTo(config.tokenEndpoint);
+        assertThat(deserialized.endSessionEndpoint).isNull();
         assertThat(deserialized.registrationEndpoint).isNull();
     }
 
@@ -167,6 +201,7 @@ public class AuthorizationServiceConfigurationTest {
         assertEquals(TEST_AUTH_ENDPOINT, config.authorizationEndpoint.toString());
         assertEquals(TEST_TOKEN_ENDPOINT, config.tokenEndpoint.toString());
         assertEquals(TEST_REGISTRATION_ENDPOINT, config.registrationEndpoint.toString());
+        assertEquals(TEST_END_SESSION_ENDPOINT, config.endSessionEndpoint.toString());
     }
 
     @Test
@@ -196,7 +231,8 @@ public class AuthorizationServiceConfigurationTest {
         InputStream is = new ByteArrayInputStream(TEST_JSON.getBytes());
         when(mHttpConnection.getInputStream()).thenReturn(is);
         doFetch();
-        mCallback.waitForCallback();
+        mPausedExecutorService.runAll();
+        shadowOf(getMainLooper()).idle();
         AuthorizationServiceConfiguration result = mCallback.config;
         assertNotNull(result);
         assertEquals(TEST_AUTH_ENDPOINT, result.authorizationEndpoint.toString());
@@ -209,7 +245,9 @@ public class AuthorizationServiceConfigurationTest {
         InputStream is = new ByteArrayInputStream(TEST_JSON.getBytes());
         when(mHttpConnection.getInputStream()).thenReturn(is);
         doFetch();
-        mCallback.waitForCallback();
+        shadowOf(getMainLooper()).idle();
+        mPausedExecutorService.runAll();
+        shadowOf(getMainLooper()).idle();
         AuthorizationServiceConfiguration result = mCallback.config;
         assertNotNull(result);
         assertEquals(TEST_AUTH_ENDPOINT, result.authorizationEndpoint.toString());
@@ -222,7 +260,8 @@ public class AuthorizationServiceConfigurationTest {
         InputStream is = new ByteArrayInputStream(TEST_JSON_MISSING_ARGUMENT.getBytes());
         when(mHttpConnection.getInputStream()).thenReturn(is);
         doFetch();
-        mCallback.waitForCallback();
+        mPausedExecutorService.runAll();
+        shadowOf(getMainLooper()).idle();
         assertNotNull(mCallback.error);
         assertEquals(GeneralErrors.INVALID_DISCOVERY_DOCUMENT, mCallback.error);
     }
@@ -232,7 +271,8 @@ public class AuthorizationServiceConfigurationTest {
         InputStream is = new ByteArrayInputStream(TEST_JSON_MALFORMED.getBytes());
         when(mHttpConnection.getInputStream()).thenReturn(is);
         doFetch();
-        mCallback.waitForCallback();
+        mPausedExecutorService.runAll();
+        shadowOf(getMainLooper()).idle();
         assertNotNull(mCallback.error);
         assertEquals(GeneralErrors.JSON_DESERIALIZATION_ERROR, mCallback.error);
     }
@@ -242,7 +282,8 @@ public class AuthorizationServiceConfigurationTest {
         IOException ex = new IOException();
         when(mHttpConnection.getInputStream()).thenThrow(ex);
         doFetch();
-        mCallback.waitForCallback();
+        mPausedExecutorService.runAll();
+        shadowOf(getMainLooper()).idle();
         assertNotNull(mCallback.error);
         assertEquals(GeneralErrors.NETWORK_ERROR, mCallback.error);
     }
@@ -256,14 +297,8 @@ public class AuthorizationServiceConfigurationTest {
 
     private static class RetrievalCallback implements
             AuthorizationServiceConfiguration.RetrieveConfigurationCallback {
-        private Semaphore mSemaphore = new Semaphore(0);
         public AuthorizationServiceConfiguration config;
         public AuthorizationException error;
-
-        public void waitForCallback() throws Exception {
-            assertTrue(mSemaphore.tryAcquire(CALLBACK_TIMEOUT_MILLIS,
-                    TimeUnit.MILLISECONDS));
-        }
 
         @Override
         public void onFetchConfigurationCompleted(
@@ -272,7 +307,6 @@ public class AuthorizationServiceConfigurationTest {
             assertTrue((serviceConfiguration == null) ^ (ex == null));
             this.config = serviceConfiguration;
             this.error = ex;
-            mSemaphore.release();
         }
     }
 
