@@ -19,23 +19,22 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.text.TextUtils;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import net.openid.appauth.connectivity.ConnectionBuilder;
 import net.openid.appauth.connectivity.DefaultConnectionBuilder;
-
 import okio.Buffer;
 import okio.BufferedSource;
 import okio.Okio;
-
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.IOException;
 import java.lang.ref.WeakReference;
 import java.nio.charset.Charset;
+
 
 /**
  * Reads and validates the demo app configuration from `res/raw/auth_config.json`. Configuration
@@ -62,9 +61,11 @@ public final class Configuration {
     private String mClientId;
     private String mScope;
     private Uri mRedirectUri;
+    private Uri mEndSessionRedirectUri;
     private Uri mDiscoveryUri;
     private Uri mAuthEndpointUri;
     private Uri mTokenEndpointUri;
+    private Uri mEndSessionEndpoint;
     private Uri mRegistrationEndpointUri;
     private Uri mUserInfoEndpointUri;
     private boolean mHttpsRequired;
@@ -143,6 +144,9 @@ public final class Configuration {
     }
 
     @Nullable
+    public Uri getEndSessionRedirectUri() { return mEndSessionRedirectUri; }
+
+    @Nullable
     public Uri getAuthEndpointUri() {
         return mAuthEndpointUri;
     }
@@ -150,6 +154,11 @@ public final class Configuration {
     @Nullable
     public Uri getTokenEndpointUri() {
         return mTokenEndpointUri;
+    }
+
+    @Nullable
+    public Uri getEndSessionEndpoint() {
+        return mEndSessionEndpoint;
     }
 
     @Nullable
@@ -196,6 +205,7 @@ public final class Configuration {
         mClientId = getConfigString("client_id");
         mScope = getRequiredConfigString("authorization_scope");
         mRedirectUri = getRequiredConfigUri("redirect_uri");
+        mEndSessionRedirectUri = getRequiredConfigUri("end_session_redirect_uri");
 
         if (!isRedirectUriRegistered()) {
             throw new InvalidConfigurationException(
@@ -210,6 +220,7 @@ public final class Configuration {
 
             mTokenEndpointUri = getRequiredConfigWebUri("token_endpoint_uri");
             mUserInfoEndpointUri = getRequiredConfigWebUri("user_info_endpoint_uri");
+            mEndSessionEndpoint = getRequiredConfigUri("end_session_endpoint");
 
             if (mClientId == null) {
                 mRegistrationEndpointUri = getRequiredConfigWebUri("registration_endpoint_uri");
