@@ -16,8 +16,8 @@ package net.openid.appauth.browser;
 
 import android.content.pm.PackageInfo;
 import android.content.pm.Signature;
-import android.support.annotation.NonNull;
 import android.util.Base64;
+import androidx.annotation.NonNull;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
