@@ -16,11 +16,11 @@ package net.openid.appauth.internal;
 
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.customtabs.CustomTabsService;
-import android.support.v4.util.Pair;
 import android.text.TextUtils;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.browser.customtabs.CustomTabsService;
+import androidx.core.util.Pair;
 
 import net.openid.appauth.Preconditions;
 
