@@ -28,7 +28,7 @@ import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk=16)
+@Config(sdk=16)
 public final class AsciiStringListUtilTest {
 
     private static final List<String> SCOPES = Arrays.asList("email", "profile", "openid");
