@@ -16,14 +16,17 @@ package net.openid.appauth;
 
 import static android.app.Activity.RESULT_CANCELED;
 import static android.app.Activity.RESULT_OK;
-import static org.assertj.android.api.Assertions.assertThat;
+import static androidx.test.ext.truth.content.IntentSubject.assertThat;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.robolectric.Shadows.shadowOf;
 
 import android.app.PendingIntent;
+import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+
+import androidx.test.core.app.ApplicationProvider;
 
 import net.openid.appauth.AuthorizationException.AuthorizationRequestErrors;
 import org.junit.Before;
@@ -31,52 +34,72 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.RuntimeEnvironment;
 import org.robolectric.android.controller.ActivityController;
 import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowActivity;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk=16)
+@Config(sdk = 16)
 public class AuthorizationManagementActivityTest {
 
+    private Context mContext;
     private AuthorizationRequest mAuthRequest;
+    private EndSessionRequest mEndSessionRequest;
     private Intent mAuthIntent;
     private Intent mCompleteIntent;
     private PendingIntent mCompletePendingIntent;
     private Intent mCancelIntent;
     private PendingIntent mCancelPendingIntent;
-    private Intent mStartIntentWithPendings;
-    private Intent mStartIntentWithPendingsWithoutCancel;
-    private Intent mStartForResultIntent;
+    private Intent mStartAuthIntentWithPendings;
+    private Intent mStartAuthIntentWithPendingsWithoutCancel;
+    private Intent mStartAuthForResultIntent;
+    private Intent mEndSessionIntentWithPendings;
+    private Intent mEndSessionIntentWithPendingsWithoutCancel;
+    private Intent mEndSessionForResultIntent;
     private ActivityController<AuthorizationManagementActivity> mController;
     private AuthorizationManagementActivity mActivity;
     private ShadowActivity mActivityShadow;
     private Uri mSuccessAuthRedirect;
     private Uri mErrorAuthRedirect;
+    private Uri mSuccessEndSessionRedirect;
+    private Uri mErrorEndSessionRedirect;
 
     @Before
     public void setUp() {
+        mContext = ApplicationProvider.getApplicationContext();
+
         mAuthRequest = TestValues.getTestAuthRequest();
+        mEndSessionRequest = TestValues.getTestEndSessionRequest();
 
         mAuthIntent = new Intent("AUTH");
         mCompleteIntent = new Intent("COMPLETE");
         mCompletePendingIntent =
-                PendingIntent.getActivity(RuntimeEnvironment.application, 0, mCompleteIntent, 0);
+                PendingIntent.getActivity(mContext, 0, mCompleteIntent, 0);
         mCancelIntent = new Intent("CANCEL");
         mCancelPendingIntent =
-                PendingIntent.getActivity(RuntimeEnvironment.application, 0, mCancelIntent, 0);
+                PendingIntent.getActivity(mContext, 0, mCancelIntent, 0);
 
-        mStartIntentWithPendings =
+        mStartAuthIntentWithPendings =
                 createStartIntentWithPendingIntents(mAuthRequest, mCancelPendingIntent);
-        mStartIntentWithPendingsWithoutCancel =
+        mStartAuthIntentWithPendingsWithoutCancel =
                 createStartIntentWithPendingIntents(mAuthRequest, null);
-        mStartForResultIntent = createStartForResultIntent(mAuthRequest);
+        mStartAuthForResultIntent = createStartForResultIntent(mAuthRequest);
+
+        mEndSessionIntentWithPendings =
+            createStartIntentWithPendingIntents(mEndSessionRequest, mCancelPendingIntent);
+        mEndSessionIntentWithPendingsWithoutCancel =
+            createStartIntentWithPendingIntents(mEndSessionRequest, null);
+        mEndSessionForResultIntent =
+            createStartForResultIntent(mEndSessionRequest);
 
         mSuccessAuthRedirect = mAuthRequest.redirectUri.buildUpon()
-                .appendQueryParameter(AuthorizationResponse.KEY_STATE, mAuthRequest.state)
+                .appendQueryParameter(AuthorizationResponse.KEY_STATE, mAuthRequest.getState())
                 .appendQueryParameter(AuthorizationResponse.KEY_AUTHORIZATION_CODE, "12345")
                 .build();
+
+        mSuccessEndSessionRedirect = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(EndSessionRequest.KEY_STATE, mEndSessionRequest.getState())
+            .build();
 
         mErrorAuthRedirect = mAuthRequest.redirectUri.buildUpon()
                 .appendQueryParameter(
@@ -87,14 +110,22 @@ public class AuthorizationManagementActivityTest {
                         AuthorizationRequestErrors.ACCESS_DENIED.errorDescription)
                 .build();
 
-        instantiateActivity(mStartIntentWithPendings);
+        mErrorEndSessionRedirect = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(AuthorizationException.PARAM_ERROR,
+                AuthorizationRequestErrors.ACCESS_DENIED.error)
+            .appendQueryParameter(
+                AuthorizationException.PARAM_ERROR_DESCRIPTION,
+                AuthorizationRequestErrors.ACCESS_DENIED.errorDescription)
+            .build();
+
+        instantiateActivity(mStartAuthIntentWithPendings);
     }
 
     private Intent createStartIntentWithPendingIntents(
-            AuthorizationRequest authRequest,
+            AuthorizationManagementRequest authRequest,
             PendingIntent cancelIntent) {
         return AuthorizationManagementActivity.createStartIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 authRequest,
                 mAuthIntent,
                 mCompletePendingIntent,
@@ -102,9 +133,9 @@ public class AuthorizationManagementActivityTest {
     }
 
     private Intent createStartForResultIntent(
-            AuthorizationRequest authRequest) {
+        AuthorizationManagementRequest authRequest) {
         return AuthorizationManagementActivity.createStartForResultIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 authRequest,
                 mAuthIntent);
     }
@@ -119,9 +150,9 @@ public class AuthorizationManagementActivityTest {
     }
 
     @Test
-    public void testSuccessFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
+    public void testLoginSuccessFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
         // start the flow
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -134,24 +165,24 @@ public class AuthorizationManagementActivityTest {
         // on completion of the authorization activity, the result will be forwarded to the
         // management activity via newIntent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 mSuccessAuthRedirect));
 
         // the management activity is then resumed
         mController.resume();
 
         // after which the completion intent should be fired
-        assertThat(mActivityShadow.getNextStartedActivity())
-                .hasAction("COMPLETE")
-                .hasData(mSuccessAuthRedirect)
-                .hasExtra(AuthorizationResponse.EXTRA_RESPONSE);
-        assertThat(mActivity).isFinishing();
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mSuccessAuthRedirect);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationResponse.EXTRA_RESPONSE);
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testSuccessFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
+    public void testEndOfSessionSuccessFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
         // start the flow
-        instantiateActivity(mStartForResultIntent);
+        instantiateActivity(mEndSessionIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -164,7 +195,36 @@ public class AuthorizationManagementActivityTest {
         // on completion of the authorization activity, the result will be forwarded to the
         // management activity via newIntent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+            mContext,
+            mSuccessEndSessionRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // after which the completion intent should be fired
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mSuccessEndSessionRedirect);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginSuccessFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
+        // start the flow
+        instantiateActivity(mStartAuthForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // the management activity will be paused while the authorization flow is running.
+        // if there is no memory pressure, the activity will remain in the paused state.
+        mController.pause();
+
+        // on completion of the authorization activity, the result will be forwarded to the
+        // management activity via newIntent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+                mContext,
                 mSuccessAuthRedirect));
 
         // the management activity is then resumed
@@ -172,13 +232,40 @@ public class AuthorizationManagementActivityTest {
 
         // and then sets a result before finishing as there is no completion intent
         assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testSuccessFlow_withPendingIntentsAndWithDestroy_shouldSendCompleteIntent() {
+    public void testEndSessionSuccessFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
         // start the flow
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mEndSessionForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // the management activity will be paused while the authorization flow is running.
+        // if there is no memory pressure, the activity will remain in the paused state.
+        mController.pause();
+
+        // on completion of the authorization activity, the result will be forwarded to the
+        // management activity via newIntent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+            mContext,
+            mSuccessEndSessionRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // and then sets a result before finishing as there is no completion intent
+        assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginSuccessFlow_withPendingIntentsAndWithDestroy_shouldSendCompleteIntent() {
+        // start the flow
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -190,29 +277,29 @@ public class AuthorizationManagementActivityTest {
         mController.pause().stop().saveInstanceState(savedState).destroy();
 
         // on completion of the authorization activity, a new management activity will be created
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create(savedState).start();
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 mSuccessAuthRedirect));
 
         // the management activity is then resumed
         mController.resume();
 
         // after which the completion intent should be fired
-        assertThat(mActivityShadow.getNextStartedActivity())
-                .hasAction("COMPLETE")
-                .hasData(mSuccessAuthRedirect)
-                .hasExtra(AuthorizationResponse.EXTRA_RESPONSE);
-        assertThat(mActivity).isFinishing();
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mSuccessAuthRedirect);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationResponse.EXTRA_RESPONSE);
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testSuccessFlow_withoutPendingIntentsAndWithDestroy_shouldSetResult() {
+    public void testEndSessionSuccessFlow_withPendingIntentsAndWithDestroy_shouldSendCompleteIntent() {
         // start the flow
-        instantiateActivity(mStartForResultIntent);
+        instantiateActivity(mEndSessionIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -224,12 +311,45 @@ public class AuthorizationManagementActivityTest {
         mController.pause().stop().saveInstanceState(savedState).destroy();
 
         // on completion of the authorization activity, a new management activity will be created
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create(savedState).start();
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+            mContext,
+            mSuccessAuthRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // after which the completion intent should be fired
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mSuccessAuthRedirect);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginSuccessFlow_withoutPendingIntentsAndWithDestroy_shouldSetResult() {
+        // start the flow
+        instantiateActivity(mStartAuthForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // on a device under memory pressure, the management activity will be destroyed, but
+        // it will be able to save its state
+        Bundle savedState = new Bundle();
+        mController.pause().stop().saveInstanceState(savedState).destroy();
+
+        // on completion of the authorization activity, a new management activity will be created
+        instantiateActivity(mStartAuthIntentWithPendings);
+        mController.create(savedState).start();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+                mContext,
                 mSuccessAuthRedirect));
 
         // the management activity is then resumed
@@ -237,13 +357,44 @@ public class AuthorizationManagementActivityTest {
 
         // and then sets a result before finishing as there is no completion intent
         assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testFailureFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
+    public void testEndSessionSuccessFlow_withoutPendingIntentsAndWithDestroy_shouldSetResult() {
         // start the flow
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mEndSessionForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // on a device under memory pressure, the management activity will be destroyed, but
+        // it will be able to save its state
+        Bundle savedState = new Bundle();
+        mController.pause().stop().saveInstanceState(savedState).destroy();
+
+        // on completion of the authorization activity, a new management activity will be created
+        instantiateActivity(mStartAuthIntentWithPendings);
+        mController.create(savedState).start();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+            mContext,
+            mSuccessEndSessionRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // and then sets a result before finishing as there is no completion intent
+        assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginFailureFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
+        // start the flow
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -255,24 +406,24 @@ public class AuthorizationManagementActivityTest {
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 mErrorAuthRedirect));
 
         // the management activity is then resumed
         mController.resume();
 
         // after which the completion intent should be fired
-        assertThat(mActivityShadow.getNextStartedActivity())
-                .hasAction("COMPLETE")
-                .hasData(mErrorAuthRedirect)
-                .hasExtra(AuthorizationException.EXTRA_EXCEPTION);
-        assertThat(mActivity).isFinishing();
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mErrorAuthRedirect);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testFailureFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
+    public void testEndSessionFailureFlow_withPendingIntentsAndWithoutDestroy_shouldSendCompleteIntent() {
         // start the flow
-        instantiateActivity(mStartForResultIntent);
+        instantiateActivity(mEndSessionIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -284,7 +435,36 @@ public class AuthorizationManagementActivityTest {
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+            mContext,
+            mErrorEndSessionRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // after which the completion intent should be fired
+        Intent nextStartedActivity = mActivityShadow.getNextStartedActivity();
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(mErrorEndSessionRedirect);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginFailureFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
+        // start the flow
+        instantiateActivity(mStartAuthForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // the management activity will be paused while the authorization flow is running.
+        // if there is no memory pressure, the activity will remain in the paused state.
+        mController.pause();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+                mContext,
                 mErrorAuthRedirect));
 
         // the management activity is then resumed
@@ -294,12 +474,40 @@ public class AuthorizationManagementActivityTest {
         assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
         Intent resultIntent = mActivityShadow.getResultIntent();
         assertThat(resultIntent.getData()).isEqualTo(mErrorAuthRedirect);
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testMismatchedState_withPendingIntentsAndResponseDiffersFromRequest() {
-        emulateFlowToAuthorizationActivityLaunch(mStartIntentWithPendings);
+    public void testEndSessionFailureFlow_withoutPendingIntentsAndWithoutDestroy_shouldSetResult() {
+        // start the flow
+        instantiateActivity(mEndSessionForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("AUTH");
+
+        // the management activity will be paused while the authorization flow is running.
+        // if there is no memory pressure, the activity will remain in the paused state.
+        mController.pause();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+            mContext,
+            mErrorAuthRedirect));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // after which the completion intent should be fired
+        assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_OK);
+        Intent resultIntent = mActivityShadow.getResultIntent();
+        assertThat(resultIntent.getData()).isEqualTo(mErrorEndSessionRedirect);
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginMismatchedState_withPendingIntentsAndResponseDiffersFromRequest() {
+        emulateFlowToAuthorizationActivityLaunch(mStartAuthIntentWithPendings);
 
         Uri authResponseUri = mAuthRequest.redirectUri.buildUpon()
                 .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
@@ -308,22 +516,43 @@ public class AuthorizationManagementActivityTest {
 
         Intent nextStartedActivity = emulateAuthorizationResponseReceived(
                 AuthorizationManagementActivity.createResponseHandlingIntent(
-                        RuntimeEnvironment.application,
+                        mContext,
                         authResponseUri));
 
         // the next activity should be from the completion intent, carrying an error
-        assertThat(nextStartedActivity)
-                .hasAction("COMPLETE")
-                .hasData(authResponseUri)
-                .hasExtra(AuthorizationException.EXTRA_EXCEPTION);
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(authResponseUri);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
 
         assertThat(AuthorizationException.fromIntent(nextStartedActivity))
                 .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
     }
 
     @Test
-    public void testMismatchedState_withoutPendingIntentsAndResponseDiffersFromRequest() {
-        emulateFlowToAuthorizationActivityLaunch(mStartForResultIntent);
+    public void testEndSessionnMismatchedState_withPendingIntentsAndResponseDiffersFromRequest() {
+        emulateFlowToAuthorizationActivityLaunch(mEndSessionIntentWithPendings);
+
+        Uri authResponseUri = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(EndSessionRequest.KEY_STATE, "differentState")
+            .build();
+
+        Intent nextStartedActivity = emulateAuthorizationResponseReceived(
+            AuthorizationManagementActivity.createResponseHandlingIntent(
+                mContext,
+                authResponseUri));
+
+        // the next activity should be from the completion intent, carrying an error
+        assertThat(nextStartedActivity).hasAction("COMPLETE");
+        assertThat(nextStartedActivity).hasData(authResponseUri);
+        assertThat(nextStartedActivity).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
+
+        assertThat(AuthorizationException.fromIntent(nextStartedActivity))
+            .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
+    }
+
+    @Test
+    public void testLoginMismatchedState_withoutPendingIntentsAndResponseDiffersFromRequest() {
+        emulateFlowToAuthorizationActivityLaunch(mStartAuthForResultIntent);
 
         Uri authResponseUri = mAuthRequest.redirectUri.buildUpon()
                 .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
@@ -332,7 +561,7 @@ public class AuthorizationManagementActivityTest {
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 authResponseUri));
 
         // the management activity is then resumed
@@ -342,16 +571,42 @@ public class AuthorizationManagementActivityTest {
         // via the result intent supplied to setResult
         Intent resultIntent = mActivityShadow.getResultIntent();
 
-        assertThat(resultIntent)
-                .hasData(authResponseUri)
-                .hasExtra(AuthorizationException.EXTRA_EXCEPTION);
+        assertThat(resultIntent).hasData(authResponseUri);
+        assertThat(resultIntent).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
 
         assertThat(AuthorizationException.fromIntent(resultIntent))
                 .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
     }
 
     @Test
-    public void testMismatchedState_withPendingIntentsAndNoStateInRequestWithStateInResponse() {
+    public void testEndSessionMismatchedState_withoutPendingIntentsAndResponseDiffersFromRequest() {
+        emulateFlowToAuthorizationActivityLaunch(mEndSessionForResultIntent);
+
+        Uri authResponseUri = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
+            .build();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+            mContext,
+            authResponseUri));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        // no completion intent, so exception should be passed to calling activity
+        // via the result intent supplied to setResult
+        Intent resultIntent = mActivityShadow.getResultIntent();
+
+        assertThat(resultIntent).hasData(authResponseUri);
+        assertThat(resultIntent).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
+
+        assertThat(AuthorizationException.fromIntent(resultIntent))
+            .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
+    }
+
+    @Test
+    public void testLoginMismatchedState_withPendingIntentsAndNoStateInRequestWithStateInResponse() {
         AuthorizationRequest request = new AuthorizationRequest.Builder(
                 TestValues.getTestServiceConfig(),
                 TestValues.TEST_CLIENT_ID,
@@ -371,7 +626,7 @@ public class AuthorizationManagementActivityTest {
         // the next activity should be from the completion intent, carrying an error
         Intent nextStartedActivity = emulateAuthorizationResponseReceived(
                 AuthorizationManagementActivity.createResponseHandlingIntent(
-                        RuntimeEnvironment.application,
+                        mContext,
                         authResponseUri));
 
         assertThat(AuthorizationException.fromIntent(nextStartedActivity))
@@ -379,7 +634,33 @@ public class AuthorizationManagementActivityTest {
     }
 
     @Test
-    public void testMismatchedState_withoutPendingIntentsAndNoStateInRequestWithStateInResponse() {
+    public void testEndSessionMismatchedState_withPendingIntentsAndNoStateInRequestWithStateInResponse() {
+        EndSessionRequest request = new EndSessionRequest.Builder(
+            TestValues.getTestServiceConfig(),
+            TestValues.TEST_ID_TOKEN,
+            TestValues.TEST_APP_REDIRECT_URI)
+            .setState("state")
+            .build();
+
+        Intent startIntent = createStartIntentWithPendingIntents(request, mCancelPendingIntent);
+        emulateFlowToAuthorizationActivityLaunch(startIntent);
+
+        Uri authResponseUri = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
+            .build();
+
+        // the next activity should be from the completion intent, carrying an error
+        Intent nextStartedActivity = emulateAuthorizationResponseReceived(
+            AuthorizationManagementActivity.createResponseHandlingIntent(
+                mContext,
+                authResponseUri));
+
+        assertThat(AuthorizationException.fromIntent(nextStartedActivity))
+            .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
+    }
+
+    @Test
+    public void testLoginMismatchedState_withoutPendingIntentsAndNoStateInRequestWithStateInResponse() {
         AuthorizationRequest request = new AuthorizationRequest.Builder(
                 TestValues.getTestServiceConfig(),
                 TestValues.TEST_CLIENT_ID,
@@ -398,7 +679,7 @@ public class AuthorizationManagementActivityTest {
 
         // the authorization redirect will be forwarded via a new intent
         mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
-                RuntimeEnvironment.application,
+                mContext,
                 authResponseUri));
 
         // the management activity is then resumed
@@ -410,19 +691,38 @@ public class AuthorizationManagementActivityTest {
     }
 
     @Test
-    public void testInvalidResponse() {
-        emulateFlowToAuthorizationActivityLaunch(mStartIntentWithPendings);
+    public void testEndSessionMismatchedState_withoutPendingIntentsAndNoStateInRequestWithStateInResponse() {
+        EndSessionRequest request = new EndSessionRequest.Builder(
+            TestValues.getTestServiceConfig(),
+            TestValues.TEST_ID_TOKEN,
+            TestValues.TEST_APP_REDIRECT_URI)
+            .setState("state")
+            .build();
 
-        Uri authResponseUri = mAuthRequest.redirectUri.buildUpon()
-                .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
-                .appendQueryParameter(AuthorizationResponse.KEY_AUTHORIZATION_CODE, "12345")
-                .build();
+        Intent startIntent = createStartForResultIntent(request);
+        emulateFlowToAuthorizationActivityLaunch(startIntent);
+
+        Uri authResponseUri = mEndSessionRequest.redirectUri.buildUpon()
+            .appendQueryParameter(AuthorizationResponse.KEY_STATE, "differentState")
+            .build();
+
+        // the authorization redirect will be forwarded via a new intent
+        mController.newIntent(AuthorizationManagementActivity.createResponseHandlingIntent(
+            mContext,
+            authResponseUri));
+
+        // the management activity is then resumed
+        mController.resume();
+
+        Intent resultIntent = mActivityShadow.getResultIntent();
+        assertThat(AuthorizationException.fromIntent(resultIntent))
+            .isEqualTo(AuthorizationRequestErrors.STATE_MISMATCH);
     }
 
     @Test
-    public void testCancelFlow_withPendingIntentsAndWithoutDestroy() {
+    public void testLoginCancelFlow_withPendingIntentsAndWithoutDestroy() {
         // start the flow
-        instantiateActivity(mStartIntentWithPendings);
+        instantiateActivity(mStartAuthIntentWithPendings);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -436,13 +736,33 @@ public class AuthorizationManagementActivityTest {
 
         // at which point the cancel intent should be fired
         assertThat(mActivityShadow.getNextStartedActivity()).hasAction("CANCEL");
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testCancelFlow_withoutPendingIntentsAndWithoutDestroy() {
+    public void testEndSessionCancelFlow_withPendingIntentsAndWithoutDestroy() {
         // start the flow
-        instantiateActivity(mStartForResultIntent);
+        instantiateActivity(mEndSessionIntentWithPendings);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).isNotNull();
+
+        // the management activity will be paused while this auth intent is running
+        mController.pause();
+
+        // when the user cancels the auth intent, the management activity will be resumed
+        mController.resume();
+
+        // at which point the cancel intent should be fired
+        assertThat(mActivityShadow.getNextStartedActivity()).hasAction("CANCEL");
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginCancelFlow_withoutPendingIntentsAndWithoutDestroy() {
+        // start the flow
+        instantiateActivity(mStartAuthForResultIntent);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -458,18 +778,45 @@ public class AuthorizationManagementActivityTest {
         assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_CANCELED);
 
         Intent resultIntent = mActivityShadow.getResultIntent();
-        assertThat(resultIntent).hasExtra(AuthorizationException.EXTRA_EXCEPTION);
+        assertThat(resultIntent).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
 
         assertThat(AuthorizationException.fromIntent(resultIntent))
             .isEqualTo(AuthorizationException.GeneralErrors.USER_CANCELED_AUTH_FLOW);
 
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     @Test
-    public void testCancelFlow_withCompletionIntentButNoCancelIntent() {
+    public void testEndSessionCancelFlow_withoutPendingIntentsAndWithoutDestroy() {
         // start the flow
-        instantiateActivity(mStartIntentWithPendingsWithoutCancel);
+        instantiateActivity(mEndSessionForResultIntent);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).isNotNull();
+
+        // the management activity will be paused while this auth intent is running
+        mController.pause();
+
+        // when the user cancels the auth intent, the management activity will be resumed
+        mController.resume();
+
+        // at which point the cancel intent should be fired
+        assertThat(mActivityShadow.getResultCode()).isEqualTo(RESULT_CANCELED);
+
+        Intent resultIntent = mActivityShadow.getResultIntent();
+        assertThat(resultIntent).extras().containsKey(AuthorizationException.EXTRA_EXCEPTION);
+
+        assertThat(AuthorizationException.fromIntent(resultIntent))
+            .isEqualTo(AuthorizationException.GeneralErrors.USER_CANCELED_AUTH_FLOW);
+
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testLoginCancelFlow_withCompletionIntentButNoCancelIntent() {
+        // start the flow
+        instantiateActivity(mStartAuthIntentWithPendingsWithoutCancel);
         mController.create().start().resume();
 
         // an activity should be started for auth
@@ -483,7 +830,27 @@ public class AuthorizationManagementActivityTest {
 
         // as there is no cancel intent, the activity simply finishes
         assertThat(mActivityShadow.getNextStartedActivity()).isNull();
-        assertThat(mActivity).isFinishing();
+        assertThat(mActivity.isFinishing()).isTrue();
+    }
+
+    @Test
+    public void testEndSessionCancelFlow_withCompletionIntentButNoCancelIntent() {
+        // start the flow
+        instantiateActivity(mEndSessionIntentWithPendingsWithoutCancel);
+        mController.create().start().resume();
+
+        // an activity should be started for auth
+        assertThat(mActivityShadow.getNextStartedActivity()).isNotNull();
+
+        // the management activity will be paused while this auth intent is running
+        mController.pause();
+
+        // when the user cancels the auth intent, the management activity will be resumed
+        mController.resume();
+
+        // as there is no cancel intent, the activity simply finishes
+        assertThat(mActivityShadow.getNextStartedActivity()).isNull();
+        assertThat(mActivity.isFinishing()).isTrue();
     }
 
     private void emulateFlowToAuthorizationActivityLaunch(Intent startIntent) {
