@@ -31,13 +31,14 @@ import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk=16)
+@Config(sdk = 16)
 public class AuthorizationServiceDiscoveryTest {
     // ToDo: add more tests for remaining getters
     static final String TEST_AUTHORIZATION_ENDPOINT = "http://test.openid.com/o/oauth/auth";
     static final String TEST_TOKEN_ENDPOINT = "http://test.openid.com/o/oauth/token";
     static final String TEST_USERINFO_ENDPOINT = "http://test.openid.com/o/oauth/userinfo";
     static final String TEST_REGISTRATION_ENDPOINT = "http://test.openid.com/o/oauth/register";
+    static final String TEST_END_SESSION_ENDPOINT = "http://test.openid.com/o/oauth/logout";
     static final String TEST_JWKS_URI = "http://test.openid.com/o/oauth/jwks";
     static final List<String> TEST_RESPONSE_TYPES_SUPPORTED = Arrays.asList("code", "token");
     static final List<String> TEST_SUBJECT_TYPES_SUPPORTED = Arrays.asList("public");
@@ -53,6 +54,7 @@ public class AuthorizationServiceDiscoveryTest {
         TEST_TOKEN_ENDPOINT,
         TEST_USERINFO_ENDPOINT,
         TEST_REGISTRATION_ENDPOINT,
+        TEST_END_SESSION_ENDPOINT,
         TEST_JWKS_URI,
         TEST_RESPONSE_TYPES_SUPPORTED,
         TEST_SUBJECT_TYPES_SUPPORTED,
