@@ -27,7 +27,7 @@ import org.robolectric.annotation.Config;
 
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk=16)
+@Config(sdk = 16)
 public class ClientSecretBasicTest {
     @Test
     public void testGetRequestHeaders() {
